requirements.txt
streamlit
groq
pymupdf
python-dotenv
requests
Pillow